Separated client and server projects.
