+++ MattEland.DigitalDungeonMaster.Shared/AdventureStatus.cs
@@ -1,4 +1,4 @@
-namespace MattEland.DigitalDungeonMaster.GameManagement.Models;
+namespace MattEland.DigitalDungeonMaster.Shared;
 
 public enum AdventureStatus
 {
