+++ MattEland.DigitalDungeonMaster.WebAPI/Routes/AdventureRouteExtensions.cs
@@ -1,4 +1,3 @@
-using MattEland.DigitalDungeonMaster.GameManagement.Models;
 using MattEland.DigitalDungeonMaster.GameManagement.Services;
 using MattEland.DigitalDungeonMaster.Shared;
 using MattEland.DigitalDungeonMaster.WebAPI.Models;
