+++ MattEland.DigitalDungeonMaster.WebAPI/Services/ChatService.cs
@@ -1,7 +1,6 @@
 using System.Text;
 using MattEland.DigitalDungeonMaster.Agents.GameMaster;
 using MattEland.DigitalDungeonMaster.Agents.WorldBuilder.Models;
-using MattEland.DigitalDungeonMaster.GameManagement.Models;
 using MattEland.DigitalDungeonMaster.Services;
 using MattEland.DigitalDungeonMaster.Shared;
 using MattEland.DigitalDungeonMaster.WebAPI.Models;
