+++ MattEland.DigitalDungeonMaster/GameManagement/Services/AdventuresService.cs
@@ -1,6 +1,6 @@
 using MattEland.DigitalDungeonMaster.Agents.WorldBuilder.Models;
-using MattEland.DigitalDungeonMaster.GameManagement.Models;
 using MattEland.DigitalDungeonMaster.Services;
+using MattEland.DigitalDungeonMaster.Shared;
 using Newtonsoft.Json;
 
 namespace MattEland.DigitalDungeonMaster.GameManagement.Services;
