+++ MattEland.DigitalDungeonMaster/GameManagement/Services/RulesetService.cs
@@ -1,5 +1,5 @@
-using MattEland.DigitalDungeonMaster.GameManagement.Models;
 using MattEland.DigitalDungeonMaster.Services;
+using MattEland.DigitalDungeonMaster.Shared;
 
 namespace MattEland.DigitalDungeonMaster.GameManagement.Services;
